Restructure login endpoint
Return a generic message for invalid login. Specifying the type of
error like it was before it's a security issue because if the attacker
knows the user exists (only the password is incorrect because it was
given the message "incorrect password"), then it knows that only has to
brute-force that user for several passwords. This way, making it generic,
the attacker doesn't have the information to continue a brute-force.